Rename tests folder to __mocks__
